Create README.md
DESCRIPTION
Cardiovascular diseases are one of the leading causes of deaths globally. To identify the causes and develop a system to predict potential heart attacks in an effective manner is necessary. The data presented has all the information about relevant factors that might have an impact on cardiovascular health. The data needs to be studied in detail for further analysis.
You are required to determine and examine the factors that play a significant role in increasing the rate of heart attacks. Also, use the findings to create and predict a model.

Steps to perform:
Importing, Understanding, and Inspecting Data :
Perform preliminary data analysis and find the structure, missing values and duplicates of the data.
Remove duplicates (if any) and treat missing values using an appropriate strategy.
Obtain statistical summary of the data. Explore the measures of central tendencies and spread of the data.
Performing EDA and Modeling:
Identify categorical variable in nature.
Study the relationship of CVD across different ages.
Is it possible to detect heart attack based on outliers in resting blood pressure of the patient?
Study overall patients w.r.t . gender.
Describe the relationship between cholesterol levels and CVD.
How is peak exercising and occurrence of heart attack related?
Does thalassemia play a major cause of identifying CVD? How are the other features determining the presence of CVD?
Pair plot to understand the relationship between all the variables.
Perform logistic regression, predict the outcome for test data, and validate the results by using the confusion matrix.